Removed unnecessary files and restructured files for authentication
