--- soup-app/src/components/authentication/withAuthenticate.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-
-const withAuthenticate = PassedComponent => {
-    return class HOC extends Component {
-       render() {
-           return (
-               <div>
-                   <PassedComponent />
-               </div>
-           )
-       }
-    }
-}
-
-export default withAuthenticate
